[NFC] reorganize JLArrays package
